index::names tests: add conflicts with high stages
We add entries into the main index to correspond with the NAME entries
that we're going to test.  NAME entries store the results of conflicts
occuring with rename detection during merge, and they must correspond to
conflicts in the index.

This test was mistakenly adding regular entries.  The checkout
validation failed, since it requires NAME entries to correspond to
high-stage (conflict) entries.  Correct the test to actually create
conflicts.